Open demo plot in browser
