+++ README.md
@@ -1,0 +1,20 @@
+# Spectral Pipeline
+
+Этот проект выполняет аппроксимацию сигналов из файлов `.dat` и строит графики.
+
+## Установка
+```bash
+pip install -r requirements.txt
+```
+
+## Запуск
+```bash
+python find_freqs_and_visualize.py data
+```
+После обработки скрипт открывает интерактивный график в браузере.
+
+## Тесты
+```bash
+pytest
+```
+
