+++ requirements.txt
@@ -1,0 +1,5 @@
+numpy
+pandas
+plotly
+openpyxl
+scipy
